+++ app/models/surveys/answer_session.rb
@@ -11,7 +11,7 @@ class AnswerSession < ActiveRecord::Base
 
   # Validations
   validates_presence_of :survey_id, :encounter
-  validates_uniqueness_of :survey_id, scope: [:encounter, :user_id, :child_id]
+  validates_uniqueness_of :survey_id, scope: [:encounter, :user_id, :child_id, :deleted]
 
   # Model Methods
   def self.most_recent(survey_id, user_id)
@@ -39,6 +39,10 @@ class AnswerSession < ActiveRecord::Base
     end
 
     self[:locked]
+  end
+
+  def available_for_user_types?(user_types)
+    self.survey.survey_user_types.where(user_type: user_types).count > 0
   end
 
   def process_answer(question, response)