+++ test/controllers/account_controller_test.rb
@@ -34,15 +34,6 @@ class AccountControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should get registration about-me survey for logged in user" do
-    skip "Loading from file will be deprecated. Surveys will be built in online interface"
-    Survey.load_from_file("about-me")
-    login(users(:social))
-    get :get_started_step_three
-    assert_response :success
-  end
-
-
   test "should get user_type page for logged in user" do
     login(users(:user_1))
     get :user_type