+++ test/models/survey_user_type_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class SurveyUserTypeTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end